+++ tests/functional/test_plugins.py
@@ -21,6 +21,7 @@ from .utils.tmp import make_tmp_file
 logger = get_logger()
 
 
+@pytest.mark.flaky(max_runs=5)
 @pytest.mark.parametrize("copy_files", [[LS_PATH]], indirect=True)
 def test_network(copy_files: list[Path], chalk: Chalk):
     bin_path = copy_files[0]